# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_transport_http_admin.AuditLogListResponseDoc:
    properties:
      data:
        properties:
          results:
            items:
              $ref: '#/definitions/internal_transport_http_admin.AuditLogResponse'
            type: array
          total:
            type: integer
        type: object
      errorMsg:
        type: string
    type: object
...
  internal_transport_http_admin.ErrorDoc:
    properties:
      data: {}
      details: {}
      errorCode:
        type: string
      errorMsg:
        type: string
      requestId:
        type: string
    type: object
...
  internal_transport_http_admin.SystemEventListResponseDoc:
    properties:
      data:
        properties:
          results:
            items:
              $ref: '#/definitions/internal_transport_http_admin.SystemEventResponse'
            type: array
          total:
            type: integer
        type: object
      errorMsg:
        type: string
    type: object
...
  internal_transport_http_admin.UsageLogListResponseDoc:
    properties:
      data:
        properties:
          results:
            items:
              $ref: '#/definitions/internal_transport_http_admin.UsageLogResponse'
            type: array
          total:
            type: integer
        type: object
      errorMsg:
        type: string
    type: object
...
  internal_transport_http_admin.UserAuthEventListResponseDoc:
    properties:
      data:
        properties:
          results:
            items:
              $ref: '#/definitions/internal_transport_http_admin.AuthEventResponse'
            type: array
          total:
            type: integer
        type: object
      errorMsg:
        type: string
    type: object
...
info:
  contact: {}
  description: DEEIX Chat 后端 API 文档
  title: DEEIX Chat API
  version: 0.1.0
paths:
  /admin/audit-logs:
    get:
      consumes:
      - application/json
      description: superadmin 分页查看全量可追溯审计日志
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 搜索关键词
        in: query
        name: query
        type: string
      - description: 资源类型
        in: query
        name: resource
        type: string
      - description: 动作
        in: query
        name: action
        type: string
      - description: 操作人用户ID
        in: query
        name: actor_user_id
        type: integer
      - description: 创建时间起点(RFC3339)
        in: query
        name: created_from
        type: string
      - description: 创建时间终点(RFC3339)
        in: query
        name: created_to
        type: string
      - description: 排序方式
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_admin.AuditLogListResponseDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_admin.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 超级管理员查询审计日志
      tags:
      - admin
...
  /admin/call-logs:
    get:
      consumes:
      - application/json
      description: superadmin 分页查看全量模型调用与计费用量账本
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 搜索模型、上游、绑定编码、协议
        in: query
        name: query
        type: string
      - description: 平台模型名筛选
        in: query
        name: platform_model_name
        type: string
      - description: 计费模式筛选：free/token/call/duration/tiered
        in: query
        name: billing_mode
        type: string
      - description: 调用人用户ID
        in: query
        name: user_id
        type: integer
      - description: 创建时间起点(RFC3339)
        in: query
        name: created_from
        type: string
      - description: 创建时间终点(RFC3339)
        in: query
        name: created_to
        type: string
      - description: 排序方式
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_admin.UsageLogListResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_admin.ErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_admin.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 超级管理员查询模型调用日志
      tags:
      - admin
...
  /admin/system-events:
    get:
      consumes:
      - application/json
      description: superadmin 分页查看后台结构化系统事件
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 搜索关键词
        in: query
        name: query
        type: string
      - description: 级别
        in: query
        name: level
        type: string
      - description: 来源
        in: query
        name: source
        type: string
      - description: 事件
        in: query
        name: event
        type: string
      - description: 创建时间起点(RFC3339)
        in: query
        name: created_from
        type: string
      - description: 创建时间终点(RFC3339)
        in: query
        name: created_to
        type: string
      - description: 排序方式
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_admin.SystemEventListResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_admin.ErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_admin.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 超级管理员查询系统事件
      tags:
      - admin
  /admin/user-auth-events:
    get:
      consumes:
      - application/json
      description: superadmin 分页查询认证事件，支持 user_id/event_type/result 过滤
      parameters:
      - description: 用户ID过滤
        in: query
        name: user_id
        type: integer
      - description: 事件类型过滤
        in: query
        name: event_type
        type: string
      - description: 结果过滤(success/failure/blocked)
        in: query
        name: result
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_admin.UserAuthEventListResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_admin.ErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_admin.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 超级管理员查询用户认证事件
      tags:
      - admin
  /admin/users:
    get:
      consumes:
      - application/json
      description: superadmin 分页查看所有用户，实现账户隔离管理
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_admin.UserListResponseDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_admin.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 超级管理员查询用户
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: 创建普通用户账号，superadmin 账号仅允许系统初始化唯一实例
      parameters:
      - description: 用户参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_admin.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_admin.CreateUserResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_admin.ErrorDoc'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_transport_http_admin.ErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_admin.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 超级管理员创建用户
      tags:
      - admin
  /admin/users/{id}:
    delete:
      consumes:
      - application/json
      description: superadmin 硬删除指定普通用户及其主要用户域数据
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_admin.DeleteUserResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_admin.ErrorDoc'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_transport_http_admin.ErrorDoc'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_transport_http_admin.ErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_admin.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 超级管理员删除用户
      tags:
      - admin
    patch:
      consumes:
      - application/json
      description: superadmin 统一维护角色、状态、时区等可编辑字段
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 局部更新参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_admin.PatchUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_admin.UpdateUserStatusResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_admin.ErrorDoc'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_transport_http_admin.ErrorDoc'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_transport_http_admin.ErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_admin.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 超级管理员更新用户可编辑字段
      tags:
      - admin
  /admin/users/{id}/reset-password:
    post:
      consumes:
      - application/json
      description: superadmin 重置指定用户密码并吊销其全部会话
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 重置密码参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_admin.ResetUserPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_admin.ResetUserPasswordResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_admin.ErrorDoc'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_transport_http_admin.ErrorDoc'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_transport_http_admin.ErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_admin.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 超级管理员重置用户密码
      tags:
      - admin
  /admin/users/{id}/revoke-sessions:
    post:
      consumes:
      - application/json
      description: superadmin 吊销指定用户全部活跃会话，用于安全治理和风险控制
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_admin.RevokeUserSessionsResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_admin.ErrorDoc'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_transport_http_admin.ErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_admin.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 超级管理员吊销用户全部会话
      tags:
      - admin
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"